bot/requirements.txt
psycopg2
python-telegram-bot==13.7
python-dotenv
paramiko